Split generator tests
